core: Rename 'core' module
