+++ packages/core/src/shared/meta.rs
