+++ packages/core/src/shared/mod.rs
